+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2023-02-10-2
+
+Changes:
+
+- Course news block
+
 ## v2023-02-10-1
 
 Changes: